refactor: extract twitchChannelMembership from twitchBot
twitchBot.ts had grown to complexity 76 / 326 lines. Split channel
join/part/reconcile state into twitchChannelMembership.ts, leaving
twitchBot.ts as a thin connection-lifecycle layer (~120 lines).

A follow-up pass extracted cacheChannelUserId and fireChannelJoinedHook
helpers to eliminate two duplicated patterns inside joinTwitchChannel.
