+++ src/twitch/eventsub/twitchEventSubSubscriptions.ts
@@ -2,7 +2,7 @@ import { createLogger } from '../../shared/logger';
 import { getAllEventSubStreamers, clearStreamerToken } from '../../db';
 import type { DbStreamerEventSub, EventSubConfig } from '../../db/eventSub';
 import { getUsers } from '../twitchApi';
-import { getActiveChannels } from '../twitchBot';
+import { getActiveChannels } from '../twitchChannelMembership';
 import { normalizeTwitchChannelName } from '../twitchChannelName';
 import { createEventSubSubscription, listEventSubSubscriptions, deleteEventSubSubscription, getValidToken, TwitchAuthError } from './twitchApiEventSub';
 import {
