+++ src/twitch/twitchBot.ts
@@ -7,122 +7,25 @@ import { executeMultiCommandForTwitch } from '../commands/multiCommandHandler';
 import { executeShoutoutForTwitch } from '../commands/shoutoutHandler';
 import { executeCountdownForTwitch } from '../commands/countdownHandler';
 import { setTwitchChannel } from '../shared/statusStore';
-import { getTwitchEnabledChannels } from '../db';
 import { normalizeTwitchChannelName } from './twitchChannelName';
-import { getUsers } from './twitchApi';
-import { createMutationQueue } from '../shared/mutationQueue';
 import { createLogger } from '../shared/logger';
+import {
+  setTmiClient,
+  setConnected,
+  clearMembershipState,
+  getActiveChannels,
+  reconcileJoinedChannels,
+  initializeActiveChannels,
+} from './twitchChannelMembership';
 
 const log = createLogger('Twitch');
 
 let client: tmi.Client | null = null;
 let connected = false;
-const activeChannels = new Set<string>();
-let _onChannelJoined: ((channel: string) => void) | null = null;
-
-export function setChannelJoinedHook(fn: (channel: string) => void): void {
-  _onChannelJoined = fn;
-}
-const activeChannelUserIds = new Map<string, string>();
-const membershipMutationQueue = createMutationQueue();
-// Twitch rate-limits JOIN to 20 per 10 s (2/s). 600 ms ≈ 1.67/s, ~83% of the ceiling.
-const JOIN_THROTTLE_MS = 600;
 const TWITCH_CHAT_MESSAGE_PATTERN = /^\[#[^\]]+\] <[^>]+>: /;
-
-function normalizeChannel(channel: string): string | null {
-  return normalizeTwitchChannelName(channel);
-}
-
-function isChannelJoined(channel: string): boolean {
-  if (!client || !connected) return false;
-  return client.getChannels().some((joinedChannel) => normalizeChannel(joinedChannel) === channel);
-}
 
 function fireAndForget(promise: Promise<void>, context: string): void {
   promise.catch((err) => log.error(`${context}:`, err));
-}
-
-async function partStaleChannel(channel: string): Promise<void> {
-  if (activeChannels.has(channel)) {
-    setTwitchChannel(channel, true);
-    return;
-  }
-  if (!client || !connected || !isChannelJoined(channel)) {
-    setTwitchChannel(channel, false);
-    return;
-  }
-  await client.part(channel);
-  setTwitchChannel(channel, false);
-  log.info(`Parted stale channel after reconnect: ${channel}`);
-}
-
-async function joinMissingChannel(channel: string): Promise<void> {
-  if (!activeChannels.has(channel)) return;
-  if (!client || !connected) {
-    setTwitchChannel(channel, false);
-    return;
-  }
-  if (isChannelJoined(channel)) {
-    setTwitchChannel(channel, true);
-    return;
-  }
-  await client.join(channel);
-  await new Promise<void>((resolve) => { setTimeout(resolve, JOIN_THROTTLE_MS); });
-  setTwitchChannel(channel, true);
-  try { _onChannelJoined?.(channel); } catch (err) { log.error('Channel joined hook error:', err); }
-  log.info(`Joined queued channel after reconnect: ${channel}`);
-}
-
-async function reconcileJoinedChannels(): Promise<void> {
-  if (!client || !connected) return;
-
-  const joinedChannels = client.getChannels()
-    .map((channel) => normalizeChannel(channel))
-    .filter((channel): channel is string => channel !== null);
-  const joinedChannelSet = new Set(joinedChannels);
-
-  for (const channel of joinedChannels) {
-    try {
-      await membershipMutationQueue.run(channel, () => partStaleChannel(channel));
-    } catch (err) {
-      log.error(`Failed to part stale channel ${channel}:`, err);
-    }
-  }
-
-  for (const channel of activeChannels) {
-    if (joinedChannelSet.has(channel)) continue;
-    try {
-      await membershipMutationQueue.run(channel, () => joinMissingChannel(channel));
-    } catch (err) {
-      setTwitchChannel(channel, false);
-      log.error(`Failed to join queued channel ${channel}:`, err);
-    }
-  }
-}
-
-async function initializeActiveChannels(): Promise<void> {
-  const configuredChannels = await getTwitchEnabledChannels();
-  for (const ch of configuredChannels) {
-    const normalized = normalizeChannel(ch);
-    if (!normalized) {
-      log.error(`Skipping invalid enabled channel in DB: ${ch}`);
-      continue;
-    }
-    activeChannels.add(normalized);
-    setTwitchChannel(normalized, false);
-  }
-
-  if (activeChannels.size === 0) {
-    log.warn('No enabled Twitch channels found in DB; connecting with no joined channels.');
-    return;
-  }
-
-  try {
-    const users = await getUsers([...activeChannels]);
-    for (const u of users) activeChannelUserIds.set(u.login.toLowerCase(), u.id);
-  } catch (err) {
-    log.error('Failed to resolve channel user IDs (shared-chat dedup unavailable):', err);
-  }
 }
 
 function handleTwitchMessage(
@@ -133,9 +36,9 @@ function handleTwitchMessage(
 ): void {
   try {
     if (self) return;
-    const normalizedChannel = normalizeChannel(channel);
+    const normalizedChannel = normalizeTwitchChannelName(channel);
     if (!normalizedChannel) return;
-    if (!activeChannels.has(normalizedChannel)) return;
+    if (!getActiveChannels().has(normalizedChannel)) return;
 
     // In Twitch shared chat, source-room-id differs from room-id when a message
     // originated in a partner channel and was shared into this one. Skip it entirely
@@ -159,12 +62,13 @@ function handleTwitchMessage(
 
 function onConnected(addr: string, port: number): void {
   connected = true;
+  setConnected(true);
   log.info(`Connected to ${addr}:${port}`);
-  log.info(`Listening on: ${[...activeChannels].join(', ') || '(none)'}`);
+  log.info(`Listening on: ${[...getActiveChannels()].join(', ') || '(none)'}`);
   // Reset every activeChannels entry via setTwitchChannel to a pessimistic
   // disconnected state until reconcileJoinedChannels() asynchronously
   // rechecks the actual joined memberships and corrects the status.
-  activeChannels.forEach((ch) => { setTwitchChannel(ch, false); });
+  getActiveChannels().forEach((ch) => { setTwitchChannel(ch, false); });
   void reconcileJoinedChannels().catch((err) => {
     log.error('Failed to reconcile joined channels:', err);
   });
@@ -172,6 +76,7 @@ function onConnected(addr: string, port: number): void {
 
 function onDisconnected(reason: string): void {
   connected = false;
+  setConnected(false);
   // Clear tmi.js's confirmed-channel list so it doesn't replay those
   // channels in its auto-rejoin queue on the next connect. All joining
   // is handled by reconcileJoinedChannels after 'connected' fires.
@@ -180,7 +85,7 @@ function onDisconnected(reason: string): void {
   // channels on the next connect — all joins are handled by reconcileJoinedChannels.
   (client as any).channels = [];
   log.warn(`Disconnected: ${reason}`);
-  activeChannels.forEach((ch) => { setTwitchChannel(ch, false); });
+  getActiveChannels().forEach((ch) => { setTwitchChannel(ch, false); });
 }
 
 export async function startTwitchBot(): Promise<void> {
@@ -203,6 +108,7 @@ export async function startTwitchBot(): Promise<void> {
       secure: true,
     },
   });
+  setTmiClient(client);
 
   client.on('message', handleTwitchMessage);
   client.on('connected', onConnected);
@@ -216,104 +122,17 @@ export async function startTwitchBot(): Promise<void> {
   }
 }
 
-export async function joinTwitchChannel(channel: string): Promise<void> {
-  const normalized = normalizeChannel(channel);
-  if (!normalized) {
-    throw new Error(`[Twitch] Invalid channel name: ${channel}`);
-  }
-
-  await membershipMutationQueue.run(normalized, async () => {
-    // Check inside the mutex so no concurrent join can race between the check
-    // and the actual client.join() call.
-    if (isChannelJoined(normalized)) {
-      // Already joined — sync local tracking so status store and activeChannels
-      // agree with the live tmi.js state.
-      activeChannels.add(normalized);
-      setTwitchChannel(normalized, true);
-      return;
-    }
-
-    if (!client || !connected) {
-      // Queue the desired membership locally so reconnect reconciliation can join
-      // it once the Twitch client is available again.
-      activeChannels.add(normalized);
-      setTwitchChannel(normalized, false);
-      getUsers([normalized])
-        .then(([u]) => { if (u) activeChannelUserIds.set(normalized, u.id); })
-        .catch((err) => { log.warn(`Failed to cache user ID for channel ${normalized}:`, err); });
-      return;
-    }
-
-    activeChannels.add(normalized);
-    setTwitchChannel(normalized, false);
-    try {
-      await client.join(normalized);
-      setTwitchChannel(normalized, true);
-      getUsers([normalized])
-        .then(([u]) => { if (u) activeChannelUserIds.set(normalized, u.id); })
-        .catch((err) => { log.warn(`Failed to cache user ID for channel ${normalized}:`, err); });
-    } catch (err) {
-      // Roll back local state; reconnect reconciliation will retry via activeChannels.
-      activeChannels.delete(normalized);
-      setTwitchChannel(normalized, false);
-      log.error(`Failed to join channel ${normalized}:`, err);
-      throw err;
-    }
-    try { _onChannelJoined?.(normalized); } catch (err) { log.error('Channel joined hook error:', err); }
-  });
-}
-
 export async function sayInChannel(channel: string, message: string): Promise<void> {
-  const normalized = normalizeChannel(channel);
+  const normalized = normalizeTwitchChannelName(channel);
   if (!normalized) throw new Error(`[Twitch] Invalid channel name: ${channel}`);
   if (!client || !connected) throw new Error(`[Twitch] Cannot send message — not connected`);
   await client.say(normalized, message);
 }
 
-export function getActiveChannels(): ReadonlySet<string> {
-  return activeChannels;
-}
-
-export function getActiveChannelUserIds(): ReadonlyMap<string, string> {
-  return activeChannelUserIds;
-}
-
-export async function partTwitchChannel(channel: string): Promise<void> {
-  const normalized = normalizeChannel(channel);
-  if (!normalized) return;
-
-  await membershipMutationQueue.run(normalized, async () => {
-    if (!activeChannels.has(normalized) && !isChannelJoined(normalized)) return;
-
-    if (!client || !connected) {
-      // We remove local state immediately and let reconcileJoinedChannels() part
-      // any stale tmi.js channel memberships on the next successful connect.
-      activeChannels.delete(normalized);
-      activeChannelUserIds.delete(normalized);
-      setTwitchChannel(normalized, false);
-      return;
-    }
-
-    try {
-      activeChannels.delete(normalized);
-      activeChannelUserIds.delete(normalized);
-      setTwitchChannel(normalized, false);
-      if (isChannelJoined(normalized)) {
-        await client.part(normalized);
-      }
-    } catch (err) {
-      // Keep desired membership removed so later reconciliation can retry
-      // parting any stale runtime join without restoring admin intent here.
-      log.error(`Failed to part channel ${normalized}:`, err);
-      throw err;
-    }
-  });
-}
-
 export async function stopTwitchBot(): Promise<void> {
   connected = false;
-  activeChannels.clear();
-  activeChannelUserIds.clear();
+  setConnected(false);
+  clearMembershipState();
   if (client) {
     try {
       await client.disconnect();
@@ -321,6 +140,7 @@ export async function stopTwitchBot(): Promise<void> {
       log.warn('Error during disconnect:', err);
     }
     client = null;
+    setTmiClient(null);
   }
   log.info('Disconnected.');
 }
